Rename package to duckdb-sqlalchemy
