+++ duckdb_sqlalchemy/__init__.py
@@ -205,7 +205,7 @@ class DuckDBEngineWarning(Warning):
 
 def index_warning() -> None:
     warnings.warn(
-        "duckdb-engine doesn't yet support reflection on indices",
+        "duckdb-sqlalchemy doesn't yet support reflection on indices",
         DuckDBEngineWarning,
     )
 
@@ -303,7 +303,7 @@ class DuckDBNullType(sqltypes.NullType):
 
 class Dialect(PGDialect_psycopg2):
     name = "duckdb"
-    driver = "duckdb_engine"
+    driver = "duckdb_sqlalchemy"
     _has_events = False
     supports_statement_cache = False
     supports_comments = has_comment_support()
@@ -351,7 +351,9 @@ class Dialect(PGDialect_psycopg2):
 
         ext = {k: config.pop(k) for k in list(config) if k not in core_keys}
         if supports_user_agent:
-            user_agent = f"duckdb_engine/{__version__}(sqlalchemy/{sqlalchemy_version})"
+            user_agent = (
+                f"duckdb-sqlalchemy/{__version__}(sqlalchemy/{sqlalchemy_version})"
+            )
             if "custom_user_agent" in config:
                 user_agent = f"{user_agent} {config['custom_user_agent']}"
             config["custom_user_agent"] = user_agent
