+++ duckdb_sqlalchemy/tests/snapshots/test_datatypes/test_interval/schema.sql
