+++ duckdb_sqlalchemy/tests/test_datatypes.py
