+++ duckdb_sqlalchemy/tests/test_pandas.py
