+++ noxfile.py
@@ -84,4 +84,4 @@ def tests_core(session: nox.Session, duckdb: str, sqlalchemy: str) -> None:
 @nox.session(py=["3.11"])
 def ty(session: nox.Session) -> None:
     session.install("-e", ".[dev]")
-    session.run("ty", "check", "duckdb_engine/")
+    session.run("ty", "check", "duckdb_sqlalchemy/")
